fix: edit mobile navbar, make about animation faster, take out experience view and move blog section there, fix footer positioning
