+++ src/assets/sass/components/footer.scss
@@ -2,6 +2,7 @@
     align-items: center;
     background-color: var(--portfolio-bg);
     border-top: 1px solid var(--portfolio-white);
+    bottom: 0;
     color: var(--portfolio-white);
     display: flex;
     height: 100px;
