+++ src/assets/sass/components/navbar.scss
@@ -18,7 +18,7 @@
 
 .nav_list {
     align-items: center;
-    background-color: var(--portfolio-terminal-header);
+    background-color: var(--portfolio-bg);
     display: flex;
     flex-direction: column;
     font-size: 1.5em;
@@ -26,6 +26,7 @@
     justify-content: center;
     position: absolute;
     transform: translateX(-100%);
+    // @fix
     top: 60px;
     width: 100%;
     z-index: 10;
@@ -44,6 +45,9 @@
 
 .nav_list.open {
     transform: translateX(0);
+    border-top: 1px solid var(--portfolio-terminal-green);
+    position: fixed;
+    // @fix header gap
 }
 
 .nav_item {
@@ -75,32 +79,9 @@
 
 .nav_toggle {
     display: flex;
-    flex-direction: column;
     margin: 16px;
-
-    span {
-        background: var(--portfolio-white);
-        border-radius: 4px;
-        height: 4px;
-        margin-bottom: 6px;
-        transform-origin: 5px 0px;
-        transition: all 0.2s linear;
-        width: 30px;
-    }
 
     @include tablet-portrait {
         display: none;
     }
-}
-
-.nav_toggle.open > span {
-    transform: rotate(45deg) translate(0px, 0px);
-}
-
-.nav_toggle.open > span:nth-child(2) {
-    display: none;
-}
-
-.nav_toggle.open > span:nth-child(3) {
-    transform: rotate(-45deg) translate(-5px, 2px);
 }
