+++ src/assets/sass/components/terminal.scss
@@ -57,7 +57,7 @@
     padding: 4px 16px 32px 0px;
 }
 
-$animation-delay-increment: 3.5s;
+$animation-delay-increment: 2.3s;
 .line {
     display: flex;
     margin: 12px 0;
@@ -67,9 +67,9 @@ $animation-delay-increment: 3.5s;
     overflow: hidden;
     white-space: nowrap;
     width: 0;
-    animation: typing 3.5s steps(40, end) normal forwards, cursor 3.5s step-end;
+    animation: typing 2.3s steps(40, end) normal forwards, cursor 2.3s step-end;
 
-    @for $i from 1 through 7 {
+    @for $i from 1 through 9 {
         &:nth-child(#{$i}) {
           animation-delay: $i * $animation-delay-increment;
         }
