+++ src/assets/sass/main.scss
@@ -4,7 +4,7 @@
 @import './pages/home.scss';
 @import './pages/about.scss';
 @import './pages/error.scss';
-@import './pages/resume.scss';
+@import './pages/blog.scss';
 @import './pages/work.scss';
 
 // components styles
