+++ src/assets/sass/pages/about.scss
@@ -2,64 +2,4 @@
   @include container;
 
   height: fit-content;
-  margin-top: 3rem;
-}
-
-.hashnode__title {
-  margin-top: 2rem;
-}
-
-.hashnode__posts {
-  gap: 24px;
-  justify-content: center;
-  margin: 2rem auto;
-
-  @include tablet-portrait {
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-    width: fit-content;
-  }
-}
-
-.post__card__container {
-  border: 1px solid var(--portfolio-terminal-green);
-  border-radius: 12px;
-  box-shadow: 0 0 1rem var(--portfolio-bg);
-  background-color: var(--portfolio-bg);
-  display: grid;
-  flex-direction: column;
-  height: max-content;
-  opacity: .7;
-  padding-bottom: 20px;
-  width: 270px;
-
-  @include tablet-landscape {
-    height: 250px;
-    width: 400px;
-  }
-}
-
-.post__card__img {
-  border-top-left-radius: 12px;
-  border-top-right-radius: 12px;
-  height: 200px;
-  width: 270px;
-
-  @include tablet-landscape {
-    height: 200px;
-    width: 400px;
-  }
-}
-
-.post__card__link {
-  color: var(--portfolio-white);
-  text-align: center;
-  font-size: 1.4rem;
-  font-weight: 600;
-  transition: .5s;
-}
-
-.post__card__link:hover {
-  color: var(--portfolio-terminal-green);
-  text-decoration: none;
 }
