+++ src/assets/sass/pages/blog.scss
@@ -1,0 +1,80 @@
+// .blog__container {
+//     @include tablet-portrait {
+//         padding-bottom: 20rem;
+//     }
+// }
+
+.hashnode__title {
+    margin-top: 2rem;
+    text-align: center;
+    font-size: 2.5rem;
+    color: var(--portfolio-terminal-green);
+
+    @include mobile {
+        font-size: 2rem;
+    }
+}
+
+.hashnode__posts {
+    gap: 24px;
+    align-items: center;
+    display: flex;
+    flex-direction: column;
+    margin: 2rem auto;
+
+    @include tablet-portrait {
+        display: grid;
+        grid-template-columns: 1fr 1fr;
+        width: fit-content;
+    }
+}
+
+// cards
+.post__card__container {
+    border: 1px solid var(--portfolio-terminal-green);
+    border-radius: 12px;
+    box-shadow: 0 0 1rem var(--portfolio-bg);
+    background-color: var(--portfolio-bg);
+    display: grid;
+    flex-direction: column;
+    height: max-content;
+    opacity: .7;
+    padding-bottom: 20px;
+    width: 270px;
+
+    @include tablet-landscape {
+        height: 250px;
+        width: 400px;
+    }
+}
+
+.post__card__img {
+    border-top-left-radius: 12px;
+    border-top-right-radius: 12px;
+    height: 200px;
+    width: 270px;
+
+    @include tablet-landscape {
+        height: 200px;
+        width: 400px;
+    }
+}
+
+.post__card__link {
+    color: var(--portfolio-white);
+    text-align: center;
+    font-size: 1.2rem;
+    font-weight: 600;
+    white-space: nowrap;
+    text-decoration: none;
+    transition: .5s;
+
+    @include tablet-landscape {
+        font-size: 1.4rem;
+    }
+}
+
+.post__card__link:hover {
+    color: var(--portfolio-terminal-green);
+    text-decoration: none;
+}
