--- src/assets/sass/pages/resume.scss
@@ -1,63 +1,0 @@
-.experience {
-    @include container;
-
-    gap: 32px;
-    margin: auto;
-    width: 80%;
-
-    header {
-        align-items: flex-start;
-        display: flex;
-        justify-content: space-between;
-        margin-bottom: 4px;
-    }
-
-    h4 {
-        font-weight: 500;
-    }
-
-    footer {
-        padding: 8px;
-    }
-
-    @include mobile {
-        margin: 2rem auto;
-    }
-
-    @include tablet-portrait {
-        width: 60%;
-    }
-}
-
-.experience_year {
-    font-style: italic;
-    font-weight: 400;
-    margin-right: 12px;
-}
-
-.resume_container {
-    background-color: var(--portfolio-bg);
-    display: flex;
-    justify-content: center;
-    margin-block: 44px;
-}
-
-.resume_link {
-    align-items: center;
-    color: var(--portfolio-terminal-green);
-    border-radius: 20px;
-    display: flex;
-    flex-direction: column;
-    font-weight: 500;
-    height: 150px;
-    justify-content: center;
-    margin: 10px;
-    padding: 10px;
-    text-decoration: none;
-    width: 250px;
-}
-
-.resume_link:hover {
-    border: 1px solid var(--portfolio-terminal-green);
-    transition: .5s;
-}
