+++ src/components/terminal/about_text.js
@@ -1,9 +1,11 @@
 export const about_text = [
-  "Hey there! I'm Francisco, a web developer from Argentina.",
-  "I started coding in 2020 but I've been using linux based",
-  "systems since 2016. I believe in using web development to",
-  "solve real-world problems and make life easier. I feel",
-  "confident with JavaScript anyways I'm willing to be",
-  "proficient in other languages. Currently, I find myself",
-  "studying IT Technician at UNNE.",
+  "Hey there! I'm Francisco, a web developer",
+  "from Argentina. I started coding in 2020 but",
+  "I've been using Linux-based systems since 2016.",
+  "I believe in using web development to solve ",
+  "real-world problems and make life easier.",
+  "I feel confident with JavaScript anyway, and I'm",
+  "willing to be proficient in other languages.",
+  "Currently, I find myself studying",
+  "IT Technician at UNNE.",
 ]
